Refactor: add defensive programming for database operations for approval of collaboration requests; docs: update read-me

+++ songmates_main/views.py
@@ -140,13 +140,18 @@ class CollabRequests(View):
     
     @method_decorator(login_required)
     def post(self, request, user_pk, *args, **kwargs):
+        # If the user has pressed the approve request button, find the user's
+        # profile, the new friend's profile, add to the many-to-many list of
+        # friends for both users and delete the collaboration request.
         if 'collab-approve' in request.POST:
             profile = Profile.objects.filter(user=request.user).first()
             new_friend = Profile.objects.filter(user=user_pk).first()
-            profile.friends.add(new_friend)
-            profile.save()
+            if new_friend:
+                profile.friends.add(new_friend)
+                profile.save()
             collab_request = CollabRequest.objects.filter(from_user=user_pk).first()
-            collab_request.delete()
+            if collab_request:
+                collab_request.delete()
             
         in_requests = CollabRequest.objects.filter(to_user=request.user)
         out_requests = CollabRequest.objects.filter(from_user=request.user)
